ADD: Split validation into validation and handle
