+++ src/tests/core/blocks/CMakeLists.txt
@@ -12,5 +12,13 @@ target_include_directories(test_validate_block PUBLIC ${INCLUDE_ALL})
 target_link_libraries(test_validate_block
   core_mine)
 
+# Test handle block
+add_executable(test_handle_block test_handle_block.c)
+target_compile_options(test_handle_block PRIVATE ${COMPILE_OPTIONS_STD})
+target_include_directories(test_handle_block PUBLIC ${INCLUDE_ALL})
+target_link_libraries(test_handle_block
+  handle_block)
+
 add_test(test_create_block ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/test_create_block)
 add_test(test_validate_block ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/test_validate_block)
+add_test(test_handle_block ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/test_handle_block)
